resolved `/api/meta` API endpoint error on specific misskey instance versions
In some versions of the Misskey instance, the `/api/meta` endpoint might not return emojis as expected or might throw an error. This would lead to the termination of the emoji fetching process. To handle this, a try-catch block was added to fetch emojis from `/api/meta` and handle potential errors. Now, in case the `/api/meta` call fails, it will fall back to fetching from `/api/emojis`. This increases the robustness of our emoji fetching process for different versions of Misskey instances.